Feature/migracion postgresql neon

+++ src/main/java/com/finmind/usuarios/repository/RolRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RolRepository extends JpaRepository<Rol, Byte> {
+public interface RolRepository extends JpaRepository<Rol, Short> {
 
     Optional<Rol> findByNombre(String nombre);
 }
